lint

+++ tests/realtikvtest/importintotest4/conflict_resolution_test.go
@@ -63,7 +63,7 @@ func (s *mockGCSSuite) getSourceRowCount(sourceContents []string) int {
 
 func (s *mockGCSSuite) getTaskByJob(jobID int64) *proto.Task {
 	s.T().Helper()
-	taskKey := importinto.TaskKey(int64(jobID))
+	taskKey := importinto.TaskKey(jobID)
 	task, err2 := s.taskMgr.GetTaskByKeyWithHistory(s.ctx, taskKey)
 	s.NoError(err2)
 	return task
